Update resend pending messages logic
