+++ src/apps/Chat/hooks/index.ts
@@ -2,6 +2,7 @@ import useActiveNetwork from 'apps/Chat/hooks/useActiveNetwork';
 import useActiveNetworkBalance from 'apps/Chat/hooks/useActiveNetworkBalance';
 import useDeliveryStatus from 'apps/Chat/hooks/useDeliveryStatus';
 import useNonContactAccounts from 'apps/Chat/hooks/useNonContactAccounts';
+import useResendPendingMessages from 'apps/Chat/hooks/useResendPendingMessages';
 import useUnreadMessages from 'apps/Chat/hooks/useUnreadMessages';
 import useUpdateNotificationCount from 'apps/Chat/hooks/useUpdateNotificationCount';
 
@@ -10,6 +11,7 @@ export {
   useActiveNetworkBalance,
   useDeliveryStatus,
   useNonContactAccounts,
+  useResendPendingMessages,
   useUnreadMessages,
   useUpdateNotificationCount,
 };
